+++ finance/finance.py
@@ -442,13 +442,6 @@ class res_company(models.Model):
                                          help=u'进项税额，是指纳税人购进货物、加工修理修配劳务、服务、无形资产或者不动产，支付或者负担的增值税额。')
     output_tax_account = fields.Many2one('finance.account', u"销项税科目", ondelete='restrict')
 
-    @api.model
-    def _get_operating_cost_account_id(self):
-        return self.env.ref('finance.small_business_chart2211001')
-
-    operating_cost_account_id = fields.Many2one('finance.account', default=_get_operating_cost_account_id,
-                                                ondelete='restrict',
-                                                string='生产费用科目', help='用在组装拆卸的费用上!')
 
 class bank_account(models.Model):
     _inherit = 'bank.account'
