+++ warehouse/__init__.py
@@ -8,3 +8,4 @@ import warehouse_move_line
 import move_matching
 import wizard
 import report
+import res_company
